requirements.txt
# pip install -r requirements.txy
mediapipe==0.8.3.1
opencv-python==4.5.1.48
pygame==2.0.1
PyOpenGL==3.1.5